+++ src/tools/interceptor_test/offboard_3d_cam_fix.py
@@ -152,7 +152,7 @@ def init_plot():
 async def main():
     # --- connect & arm ---
     drone = System()
-    await drone.connect(system_address="udp://:14540")
+    await drone.connect(system_address="udpin://0.0.0.0:14540")
     print("[INFO] Connecting to PX4 SITL...")
 
     async for health in drone.telemetry.health():